+++ src/QuickMarcEditor/QuickMarcDeriveWrapper.js
@@ -1,5 +1,6 @@
 import React, {
   useCallback,
+  useContext,
   useMemo,
   useState,
 } from 'react';
@@ -12,6 +13,7 @@ import QuickMarcEditor from './QuickMarcEditor';
 import getQuickMarcRecordStatus from './getQuickMarcRecordStatus';
 import {
   useAuthorityLinking,
+  useLccnDuplicationCheck,
   useValidation,
 } from '../hooks';
 import { QUICK_MARC_ACTIONS } from './constants';
@@ -31,7 +33,9 @@ import {
   removeEnteredDate,
   autopopulatePhysDescriptionField,
   autopopulateMaterialCharsField,
+  joinErrors,
 } from './utils';
+import { QuickMarcContext } from '../contexts';
 
 const propTypes = {
   action: PropTypes.oneOf(Object.values(QUICK_MARC_ACTIONS)).isRequired,
@@ -53,7 +57,9 @@ const QuickMarcDeriveWrapper = ({
   fixedFieldSpec,
 }) => {
   const showCallout = useShowCallout();
+  const { setValidationErrors } = useContext(QuickMarcContext);
   const { linkableBibFields, actualizeLinks, linkingRules } = useAuthorityLinking({ marcType, action });
+  const { validateLccnDuplication } = useLccnDuplicationCheck({ marcType });
   const [httpError, setHttpError] = useState(null);
 
   const validationContext = useMemo(() => ({
@@ -93,6 +99,14 @@ const QuickMarcDeriveWrapper = ({
       prepareForSubmit,
       combineSplitFields,
     )(formValues);
+
+    const lccnDuplicationError = await validateLccnDuplication(formValues);
+
+    if (lccnDuplicationError) {
+      setValidationErrors(curErrors => joinErrors(curErrors, lccnDuplicationError));
+
+      return;
+    }
 
     showCallout({ messageId: 'ui-quick-marc.record.saveNew.onSave' });
 
@@ -145,7 +159,15 @@ const QuickMarcDeriveWrapper = ({
         setHttpError(parsedError);
       });
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [onClose, showCallout, prepareForSubmit, actualizeLinks]);
+  }, [
+    onClose,
+    showCallout,
+    prepareForSubmit,
+    actualizeLinks,
+    validateLccnDuplication,
+    setValidationErrors,
+    joinErrors,
+  ]);
 
   return (
     <QuickMarcEditor