+++ src/queries/index.js
@@ -4,3 +4,4 @@ export * from './useAuthorityLinksCount';
 export * from './useLinkSuggestions';
 export * from './useMarcRecordMutation';
 export * from './useAuthorityFileNextHrid';
+export * from './useLccnDuplicateConfig';